Configurable kibana mount location for ci builds
The environment variable `KIBANA_MOUNT` can be used to specify a path to
a kibana source directory at runtime instead of using `master`. This is
intended for use with release-manager to ensure java-langserver uses the
same version of kibana that is used for the kibana build.

For https://github.com/elastic/release-manager/issues/560