Improve CI build with Docker (#52)
This commit updates the Docker configuration in order to follow
some best practices and fix some issues with the CI and Release process.

First, the Kibana git repository is not cloned anymore inside the
image. The `yarn` commands are also not anymore part of the image.
The Docker image size is now 1.34GB instead of 5.03GB.

In order to not run the container as root, the `ciagent` is created as
the user for running the build commands.
The user uid is defined as a build args to avoid permission
problems on the host and in the container when mounting the kibana repo
or the .m2 repository.

This should work out-of-the box on the CI server and locally.
It should also fixed elastic/release-manager#618