Fix user id and .m2 permission
By introducing a new CI user for running the build in #52, we
encountered two problems.

First one is that the CI user id we are creating in the Docker image
could be the same as the node user (1000), because it's a common value.
This commit fixes it by updating the node user id.

The second problem is a permission issue when mouting the .m2
folder into the container, which prevents Maven to work.
This commit fixes it by using a Docker volume for the .m2 cache.
It means that the cache can be shared between all builds while the `m2-vol`
volume is not deleted.